fix: simplify c-sync equation
The c-sync equation was slightly complicated and was trying to
incorporate the sync mode. The negative/positive v-sync level is
used for color mode detection and c-sync is also inverted dependent
on that. However, so far I didn't find hardware which would need that.
That could be even wrong, so for now I removed the mode from the c-sync
equation and it is again a simple XNOR, which hopefully will just work
with every mode, Hercules, CGA and HiRes-EGA.